requirements.txt
streamlit
python-docx
customtkinter 